standalone_scraper/requirements.txt
# Standalone Evo.com Scraper Dependencies
# Isolated from main hexar-backend requirements

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Data validation and models
pydantic==2.11.7

# Database
supabase==2.0.0

# Async support
asyncio-throttle==1.0.2

# Utilities
python-dotenv==1.0.0
colorama==0.4.6

# Progress tracking
tqdm==4.66.1

# Data handling
python-dateutil==2.8.2